refactor: move shared test utilities
