+++ internal/service/helpers_test.go
@@ -1,0 +1,44 @@
+package service_test
+
+import (
+	"errors"
+	"testing"
+
+	"github.com/jacoboneill/SecureBin/internal/db"
+)
+
+var ErrDBMock = errors.New("mock db error")
+
+type QuerierMock struct {
+	db.Querier
+	Calls int
+}
+
+type ResultMock struct {
+	rowsAffected int64
+}
+
+func (r ResultMock) LastInsertId() (int64, error) { return 0, nil }
+func (r ResultMock) RowsAffected() (int64, error) { return r.rowsAffected, nil }
+
+func AssertErrorsEqual(t testing.TB, expectedErr, capturedErr error) {
+	t.Helper()
+	if !errors.Is(capturedErr, expectedErr) {
+		if expectedErr == nil {
+			t.Errorf("expected no errors, got %v", capturedErr)
+		} else {
+			t.Errorf("expected error: %v, got %v", expectedErr, capturedErr)
+		}
+	}
+}
+
+func AssertCallCountsEqual(t testing.TB, expectedCallCount, capturedCallCount int) {
+	t.Helper()
+	if expectedCallCount != capturedCallCount {
+		var suffix string
+		if expectedCallCount > 1 {
+			suffix = "s"
+		}
+		t.Errorf("expected %d %s, got %d", expectedCallCount, suffix, capturedCallCount)
+	}
+}
