+++ internal/service/sessions_test.go
@@ -3,27 +3,12 @@ package service_test
 import (
 	"context"
 	"database/sql"
-	"errors"
 	"fmt"
 	"testing"
 
 	"github.com/jacoboneill/SecureBin/internal/db"
 	"github.com/jacoboneill/SecureBin/internal/service"
 )
-
-var ErrDBMock = errors.New("mock db error")
-
-type QuerierMock struct {
-	db.Querier
-	Calls int
-}
-
-type ResultMock struct {
-	rowsAffected int64
-}
-
-func (r ResultMock) LastInsertId() (int64, error) { return 0, nil }
-func (r ResultMock) RowsAffected() (int64, error) { return r.rowsAffected, nil }
 
 type GetSessionMock struct {
 	QuerierMock
@@ -65,28 +50,6 @@ func (m *DeleteSessionMock) DeleteSession(ctx context.Context, id string) (sql.R
 	}
 
 	return ResultMock{1}, nil
-}
-
-func AssertErrorsEqual(t testing.TB, expectedErr, capturedErr error) {
-	t.Helper()
-	if !errors.Is(capturedErr, expectedErr) {
-		if expectedErr == nil {
-			t.Errorf("expected no errors, got %v", capturedErr)
-		} else {
-			t.Errorf("expected error: %v, got %v", expectedErr, capturedErr)
-		}
-	}
-}
-
-func AssertCallCountsEqual(t testing.TB, expectedCallCount, capturedCallCount int) {
-	t.Helper()
-	if expectedCallCount != capturedCallCount {
-		var suffix string
-		if expectedCallCount > 1 {
-			suffix = "s"
-		}
-		t.Errorf("expected %d %s, got %d", expectedCallCount, suffix, capturedCallCount)
-	}
 }
 
 func TestValidateSession(t *testing.T) {
